테이블 데이터 초기화 및 카카오 API 적용

+++ src/main/java/com/delivery/monitor/domain/Deliveries.java
@@ -1,6 +1,5 @@
 package com.delivery.monitor.domain;
 
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import lombok.AllArgsConstructor;
@@ -16,7 +15,7 @@ public class Deliveries {
     private int delivery_id;
     private int order_id;
     private String delivery_status;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
     private Timestamp delivery_date;
 }
